// src/constants/enums/uploadType.ts
enum UploadType {
  COPYRIGHT_DOCUMENT = "copyright_document",
  COPYRIGHT_PROOF_DOCUMENT = "copyright_proof_document",
  COPYRIGHT_SUPPORTING_DOCUMENT = "copyright_supporting_document",
  OLD_COPYRIGHT_DOCUMENT = "old_copyright_document",
  COMPLAINT_DOCUMENT = "complaint_document",
  OWNER_IMAGE = "copyright_owners_image",
  COPYRIGHT_CLAIM_DOCUMENT = "copyright_claim_document",

  INDIVIDUAL_CITIZENSHIP = "individual_citizenship",
  INDIVIDUAL_PP = "individual_pp",

  AGENCY_PAN_VAT = "pan_vat",
  AGENCY_STAMP = "agency_stamp",
  AGENCY_REGISTRATION_DOCUMENT = "agency_registration_document",

  PROFILE_IMAGE = "profile_image",

  REMARKS_DOCUMENT = "remarks_document",
  CERTIFICATE_AUTHORITY_SIGNATURE = "certificate_authority_signature",
  STAMP = "stamp",

  COMPANY_LOGO = "company_logo",
  PAYMENT_VOUCHER = "payment_voucher",
}
export default UploadType;

// For Copyright Files.